Move cosmosdb explorer functionality out of mongo folder
