+++ src/explorer.ts
